snake head over food piece triggers action
Relates #7
