+++ index.html
@@ -14,9 +14,11 @@
         <section class="grid snake__grid">
             <div class="first"></div>
         </section>
+
+        <button class="start__btn">Start New Game</button>
     </main>
 
-    <script src="./main.js"></script>
+    <script src="./main.js" type="module"></script>
 </body>
 
 </html>
